feat: Add react-router, zustand and theming with styled components
