+++ package.json
@@ -21,8 +21,10 @@
     "lz-string": "^1.4.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.2",
     "react-scripts": "^5.0.1",
-    "styled-components": "^5.3.6"
+    "styled-components": "^5.3.6",
+    "zustand": "^4.1.1"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.5",
@@ -37,13 +39,13 @@
     "@types/styled-components": "^5.1.26",
     "@typescript-eslint/eslint-plugin": "^5.35.1",
     "@typescript-eslint/parser": "^5.35.1",
-    "gh-pages": "^4.0.0",
-    "typescript": "^4.8.3",
-    "npm-check-updates": "^16.2.1",
+    "eslint": "^8.25.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "eslint": "^8.25.0",
+    "gh-pages": "^4.0.0",
+    "npm-check-updates": "^16.2.1",
     "prettier": "^2.7.1",
+    "typescript": "^4.8.3",
     "web-vitals": "^3.0.2"
   },
   "eslintConfig": {
