+++ src/index.tsx
@@ -2,8 +2,7 @@ import ReactDOM from 'react-dom/client';
 import { StrictMode } from 'react';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-
-import './index.css';
+import { BrowserRouter } from 'react-router-dom';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement,
@@ -11,7 +10,9 @@ const root = ReactDOM.createRoot(
 
 root.render(
   <StrictMode>
-    <App />
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
   </StrictMode>,
 );
