+++ src/theme/index.ts
@@ -1,0 +1,20 @@
+/*
+  Fonts used:
+  Lilita One 400
+  Source Sans Pro 300 italic
+  Source Sans Pro 300 
+*/
+
+export const THEME = {
+  colors: {
+    white: '#ffffff',
+    black: '#000000',
+    font: '#ffffff',
+    brand: '#546785',
+    backgroundColor: '#13284a',
+  },
+  fonts: {
+    text: '"Source Sans Pro", sans-serif',
+    title: '"Lilita One", sans-serif',
+  },
+};
